Add necessary dependency
This package gets installed at some other point in the provisioning of
the development machine, as the script will run if you execute it
manually, after the machine is provisioned. This bug raises awareness
about trying to keep scripts completely self contained, which is what
you would want from a truly re-usable library of scripts.