State dependencies rather than source scripts
Sadly, the only way for that sourcing to work correctly would be for
all the scripts being referenced to be in the same working
directory that the script has been executed from. I'm not sure you could
guarantee this in the context of something like Vagrant or Packer.

Another alternative would be to bring the dependencies back in as
function calls, as per the first approach, but I prefer having the
scripts more fine grained, as it means things are a bit more composable.